+++ src/components/business/businessReview/BusinessReview.jsx
@@ -1,16 +1,16 @@
 import React from 'react'
-import BusinessHeading from "../util/BusinessHeading";
-import {useGetBusinessReviewsQuery} from "../../redux/services/businessAPI";
+import BusinessHeading from "../../util/BusinessHeading";
+import {useGetBusinessReviewsQuery} from "../../../redux/services/businessAPI";
 import {useSelector} from "react-redux";
-import GetLoader, {DISPLAY, SPINNERS} from "../util/customSpinner/GetLoader";
+import GetLoader, {DISPLAY, SPINNERS} from "../../util/customSpinner/GetLoader";
 import {Image, List, Row, Space} from "antd";
 import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
 import {faPhone, faReply} from "@fortawesome/free-solid-svg-icons";
 import useBreakpoint from "antd/es/grid/hooks/useBreakpoint";
-import {PUBLIC_BUSINESS_API} from "../../redux/api_url";
+import {PUBLIC_BUSINESS_API} from "../../../redux/api_url";
 import "./businessReview.css"
-import GetRating from "../util/GetRating";
-import NoDataGIF from "../util/NoDataGIF";
+import GetRating from "../../util/GetRating";
+import NoDataGIF from "../../util/NoDataGIF";
 
 const BASE_URL = process.env.BASE_API_URL;
