+++ src/screens/ContactRequestScreen.jsx
@@ -1,7 +1,7 @@
 import React, {useContext, useEffect} from "react";
 import {ActiveNavigationMenuContext} from "../context/ActiveNavigationProvider";
 import {NAVIGATION_ROUTES} from "../components/util/Constants";
-import ContactRequest from "../components/contactRequest/ContactRequest";
+import ContactRequest from "../components/business/contactRequest/ContactRequest";
 
 const ContactRequestScreen = () => {
     const {updateActiveNavigationMenu} = useContext(ActiveNavigationMenuContext);
