+++ src/screens/GetLocalsHome.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const GetLocalsHome = () => {
+    return <div>Get Locals</div>
+}
+
+export default GetLocalsHome;
