+++ src/screens/ReviewScreen.jsx
@@ -1,7 +1,7 @@
 import React, {useContext, useEffect} from 'react';
 import {ActiveNavigationMenuContext} from "../context/ActiveNavigationProvider";
 import {NAVIGATION_ROUTES} from "../components/util/Constants";
-import BusinessReview from "../components/businessReview/BusinessReview";
+import BusinessReview from "../components/business/businessReview/BusinessReview";
 
 export default function () {
